DHtactic_au/requirements.txt
streamlit
pandas
matplotlib
seaborn
openpyxl
Pillow
torch
torchvision